+++ src/steps/resource-manager/subscriptions/client.test.ts
@@ -7,16 +7,10 @@ import {
 } from '../../../../test/helpers/recording';
 import { J1SubscriptionClient } from './client';
 import { Location, Subscription } from '@azure/arm-subscriptions/esm/models';
-import { IntegrationConfig } from '../../../types';
-import { configFromEnv } from '../../../../test/integrationInstanceConfig';
-
-// developer used different creds than ~/test/integrationInstanceConfig
-const config: IntegrationConfig = {
-  clientId: process.env.CLIENT_ID || 'clientId',
-  clientSecret: process.env.CLIENT_SECRET || 'clientSecret',
-  directoryId: '992d7bbe-b367-459c-a10f-cf3fd16103ab',
-  subscriptionId: 'd3803fd6-2ba4-4286-80aa-f3d613ad59a7',
-};
+import {
+  config,
+  configFromEnv,
+} from '../../../../test/integrationInstanceConfig';
 
 let recording: Recording;
 
@@ -80,5 +74,65 @@ describe('iterateLocations', () => {
     });
 
     expect(locations.length).toBeGreaterThan(0);
+  });
+});
+
+describe('fetchSubscription', () => {
+  test('fetchSubscription', async () => {
+    recording = setupAzureRecording({
+      directory: __dirname,
+      name: 'fetchSubscription',
+      options: {
+        matchRequestsBy: getMatchRequestsBy({ config: configFromEnv }),
+      },
+    });
+
+    const client = new J1SubscriptionClient(
+      configFromEnv,
+      createMockIntegrationLogger(),
+    );
+
+    const subscription = await client.fetchSubscription(
+      configFromEnv.subscriptionId!,
+    );
+
+    expect(subscription).toMatchObject({
+      authorizationSource: expect.any(String),
+      displayName: expect.any(String),
+      id: expect.any(String),
+      state: expect.any(String),
+      subscriptionId: expect.any(String),
+      subscriptionPolicies: expect.any(Object),
+    });
+  });
+});
+
+describe('fetchSubscriptions', () => {
+  test('all', async () => {
+    recording = setupAzureRecording({
+      directory: __dirname,
+      name: 'fetchSubscriptions',
+      options: {
+        matchRequestsBy: getMatchRequestsBy({ config: configFromEnv }),
+      },
+    });
+
+    const client = new J1SubscriptionClient(
+      configFromEnv,
+      createMockIntegrationLogger(),
+    );
+
+    const subscriptions = await client.fetchSubscriptions();
+
+    expect(subscriptions).toContainEqual(
+      expect.objectContaining({
+        authorizationSource: expect.any(String),
+        displayName: expect.any(String),
+        id: expect.any(String),
+        state: expect.any(String),
+        subscriptionId: expect.any(String),
+        subscriptionPolicies: expect.any(Object),
+      }),
+    );
   });
 });