+++ test/integrationInstanceConfig.ts
@@ -10,7 +10,7 @@ if (process.env.LOAD_ENV) {
 
 // TODO use polly matching to find/replace configuration variables in files
 // in order to allow multiple devs to generate recordings & test files.
-const config: IntegrationConfig = {
+export const config: IntegrationConfig = {
   clientId: process.env.CLIENT_ID || 'clientId',
   clientSecret: process.env.CLIENT_SECRET || 'clientSecret',
   directoryId: 'a76fc728-0cba-45f0-a9eb-d45207e14513',